Do not delegate to helper functions in `player_input` system
Unlike Legion, Bevy requires you to add all the Queries you'll need in a
system in its function signature. This makes the use of helper
functions a little more cumbersome as it is no longer possible to just
pass the world and resources to the helpers and have them determine the
query; instead a system would need to aggregate all the queries which
might be needed and pass them on to the helpers. This is leaks
implementation details to the main system (the removal of which is the
whole point of helper functions) and adds clutter.

Instead, the helpers should be replaced by systems in their own right
and calling them is replaced by sending events. This allows the
declaration and introduction of additional queries.