Enable turn-based movement
The iyes_loopless crate is used instead of Bevy's built-in `State`
management because the latter does not work across different stages.
However, the `player_input` system needs to be in its own stage. The
crate is the reference impl for an RFC which might get merged into
future versions of Bevy anyway.